fix: Rename lexer.rs to lr_lexer.rs to avoid clippy module_inception
Clippy complained about having `src/lexer/mod.rs` declare `mod lexer;`,
which creates the pattern `lexer::lexer`. Renamed to `lr_lexer` to fix.

Fixes clippy error:
  error: module has the same name as its containing module
   --> prqlc/prqlc-parser/src/lexer/mod.rs:1:1
