+++ prqlc/prqlc-parser/src/lexer/mod.rs
@@ -1,16 +1,16 @@
-mod lexer;
+mod lr_lexer;
 
 pub mod lr;
 #[cfg(test)]
 mod test;
 
-pub use lexer::{lex_source, lex_source_recovery};
+pub use lr_lexer::{lex_source, lex_source_recovery};
 
 #[cfg(test)]
 pub mod debug {
     use super::*;
 
     pub fn lex_debug(source: &str) -> Result<lr::Tokens, Vec<crate::error::Error>> {
-        lexer::lex_source(source)
+        lr_lexer::lex_source(source)
     }
 }
